there it goes

+++ city.h
@@ -11,17 +11,23 @@ protected:
 public:
   City(const string& cName, double x, double y): coordinateX(x), coordinateY(y), name(cName){ }
   City(const City& c): coordinateX(c.coordinateX), coordinateY(c.coordinateY), name(c.name){}
-  const int operator-(const City& c) const{
-    int x = this->coordinateX - c.coordinateX;
-    int y = this->coordinateY - c.coordinateY;
+
+  //Leírhatod hogy x = City1 - City2 és visszaadja a távolságukat
+  const double operator-(const City& c) const{
+    double x = this->coordinateX - c.coordinateX;
+    double y = this->coordinateY - c.coordinateY;
     return x * x + y * y;
   }
+
+  //Megmondja melyiknek nagyobb az x coodinátája
   bool operator<(const City& c) const{
     return this->coordinateX < c.coordinateX;
   }
   bool operator>(const City& c) const{
     return this->coordinateX > c.coordinateX;
   }
+
+  //Leírthatod mondjuk, h cities[5] = City2; (City2 City típusú, cities meg City[] típusú)
   City& operator=(const City& c){
     if(&c == this) return *this;;
     this->name = c.name;
@@ -29,6 +35,9 @@ public:
     this->coordinateY = c.coordinateY;
     return *this;
   }
+
+  //Kiírja a city nevét
+  //Leírhatod h cout << City2; és kiírja
   friend std::ostream& operator<<(std::ostream& os, const City& c){
     os << c.name << std::endl;
     return os;
